Moving stat to stat.go and cleaning up some tests
